requirements.txt
APScheduler==3.0.5
fitbit==0.2.2
Flask==0.10.1
Flask-APScheduler==1.3.7
futures==3.0.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.0.3
python-dateutil==2.4.2
python-firebase==1.2
pytz==2016.3
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
tzlocal==1.2.2
Werkzeug==0.11.8
